notice_notifier: factor out #inspect and reuse in other notifiers
It's hard to inspect PerformanceNotifier and DeployNotifier because they print
too much information and take all the space in your console. We already use
custom inspect for NoticeNotifier, so we can easily reuse it there. I extracted
the code to a module and included it into 3 classes. Neat!
