+++ lib/airbrake-ruby.rb
@@ -15,6 +15,7 @@ require 'airbrake-ruby/async_sender'
 require 'airbrake-ruby/response'
 require 'airbrake-ruby/nested_exception'
 require 'airbrake-ruby/ignorable'
+require 'airbrake-ruby/inspectable'
 require 'airbrake-ruby/notice'
 require 'airbrake-ruby/backtrace'
 require 'airbrake-ruby/truncator'
