+++ lib/airbrake-ruby/performance_notifier.rb
@@ -5,6 +5,8 @@ module Airbrake
   # @api public
   # @since v3.2.0
   class PerformanceNotifier
+    include Inspectable
+
     # @param [Airbrake::Config] config
     def initialize(config)
       @config =
