+++ spec/notice_notifier_spec.rb
@@ -402,35 +402,5 @@ RSpec.describe Airbrake::NoticeNotifier do
       subject.merge_context(apples: 'oranges')
     end
   end
-
-  describe "#inspect" do
-    it "displays object information" do
-      expect(subject.inspect).to match(/
-        #<Airbrake::NoticeNotifier:0x\w+\s
-          project_id="\d+"\s
-          project_key=".+"\s
-          host="http.+"\s
-          filter_chain=\[.+\]>
-      /x)
-    end
-  end
-
-  describe "#pretty_print" do
-    it "displays object information in a beautiful way" do
-      q = PP.new
-
-      # Guarding is needed to fix JRuby failure:
-      # NoMethodError: undefined method `[]' for nil:NilClass
-      q.guard_inspect_key { subject.pretty_print(q) }
-
-      expect(q.output).to match(/
-        #<Airbrake::NoticeNotifier:0x\w+\s
-          project_id="\d+"\s
-          project_key=".+"\s
-          host="http.+"\s
-          filter_chain=\[\n\s\s
-      /x)
-    end
-  end
 end
 # rubocop:enable Layout/DotPosition
